+++ WebAPIMatricula_3C2023/API.Dto.Cal/Salida/EditarCalificacion.cs
@@ -9,9 +9,9 @@ namespace API.Dto.Calificacion.Salida
     public class EditarCalificacion : General.RespuestaAPI
     {
         public int Codigo { get; set; }
-        public int NotaProyecto { get; set; }
-        public int NotaTareas { get; set; }
-        public int NotaTrabajoCotidiano { get; set; }
+        public float NotaProyecto { get; set; }
+        public float NotaTareas { get; set; }
+        public float NotaTrabajoCotidiano { get; set; }
         public int CodigoEstudiante { get; set; }
         public int CodigoProfesor { get; set; }
         public int CodigoCurso { get; set; }
